requirements.txt
ultralytics
opencv-python
numpy
pyserial
python-dotenv
pillow
requests
